separate benchmark functions into xsum_bench.c
