+++ cmake_unofficial/CMakeLists.txt
@@ -92,6 +92,7 @@ if(XXHASH_BUILD_XXHSUM)
                         "${XXHSUM_DIR}/xsum_os_specific.c"
                         "${XXHSUM_DIR}/xsum_output.c"
                         "${XXHSUM_DIR}/xsum_sanity_check.c"
+                        "${XXHSUM_DIR}/xsum_bench.c"
                 )
   add_executable(${PROJECT_NAME}::xxhsum ALIAS xxhsum)
